Add Picasso3in1 effects.
This is actually a really cool effects library. I've just spent WAY too
long trying to make it automatically demonstrable in
Nightdriver/Mesmerizer.

It's the first effect I've tinkered with where it's entirely possible to
bring down the watchdog timer.

I've modified it to just jump between fixed settings on the primary
three effects. This commit has a lot of the experimental work cruft
that I'm about to delete because hours later, there's just no gold at
the end of that rainbow. Sure, I'd LIKE to be able to demonstrate one
line, two lines, three lines, ... but not when iterated over 33 effects
times three moddes (one is the number of circles, one is the number of,
uhm, light sabers, and is the number of LS's with heavy ghost trails -
all three in 3D. But not at the risk of bringing down the board. :-(

I couldn't find a way to support moving 33 light bars with each endpoint
moving randomly in 3d space with linear interpolations AND with ghosting
and keep it from blowing our frame budget.

Bummer.

Here's the work product. I'm going to rinse out the tears and blood in
the next (?) commit.