Replace our hand-crafted checks for bounds of leds[] with isPixelValid()
Many of our old checks didn't actually test what we were using or had
off-by-ones of their own. Someday, these can become asserts or even
removed as the callers get fixed, but so much of the SM code relied on
silent failures that it's best to mimic that while we can.