Make USE_REMOTE and USE_BUTTON more self-contained.
IMO, this change stands on its own as just good style of constraining
the code that uses remote into the files and blocks that use remote.

More motivationally, the remote code is Really Annoying to build. It's
a lot of files to compile in the hope that -ffunction-sections will
actually throw them away. Better to simply not compile them. This makes
that possible.

With this change, it's possble to lib_ignore Bounce2 and IRremoteESP8266
(and U8g2 and QRcode...) these modules, though I haven't reconciled how
to do so against our recommended document and style. As such, the change
that REALLY takes advantage of this,
+lib_ignore  = U8g2
+       Bounce2
+       IRremoteESP8266
+       QRCode
+board = esp32-s3-devkitc-1-n16r8v

is not included here. I think this change stands on its own and lays
bricks for that later change.

Tested: successful buddybuild.