Remove code setting common GPIOS to inputs when ATOMLIGHT ledstripgfx.h
If a developer put LED_PINn on pins 4, 12, 13, 14, 15, this would have ensured that no data would have ever left the chip. PIN5 would have been OK only because this code "accidientally"? missed it.

I'm highly suspicious of this code and doubt that any of us can test it and prove it's correct, we can only inspect it and reason that it's wrong.

Do you agree that it's wrong?