As taunted, major C++-ification of BlurringColors/Powder
This file's version of https://github.com/PlummersSoftwareLLC/NightDriverStrip/pull/489

Ran on Mesmerizer for 30+ minutes. Did not measure before/after frame
rates but it should help that it's using more indexed addressing modes
( mov t0, 1234($t1) ) instead of recomputing index offset addresses
(lea ...) so many times per method. A loop next now is the addition of the
size of the structure we're iterating over.

Tried to pull in the changes that happend in https://github.com/PlummersSoftwareLLC/NightDriverStrip/pull/512