Allow checking of format arguments against arguments in str_sprintf.
This detects the internal loop we previousluy fell into if we tried to
str_format("%s", String()); The test on vsnprintf() presumably couldn't
save us because our arg list had decayed to a va_list

It's bracketed to only come on if GCC, but clang is so dclose to us
it'll work there, too. Provide safe fallback for other code scanners, though.

Bonus: I've identified othr optimizations in here I can do when it's not
this time of day.