Make secrets.h more immune to changes from secrets.example.h
When new fields are added in the example file, add a test in the
file that uses the real file to be sure the fields are present. This
is to solve the case where a developer has a checked-out, working tree
with a working secrets.h, syncs to head or another branch that adds
fields in secrets, and then incurs weird compilation errors when those
new fields aren't found.

This change leads the developer to the correct file and the doc
describing it for a better experience.