Remove outdated kludge for FastLED's use of register.
Web-recreation of https://github.com/PlummersSoftwareLLC/NightDriverStrip/pull/658

FastLED used to use C style hints of 'register' that were deprecated
long ago and the keyword was removed in C++17. There was a kludge
that just defined register away.
FastLED was been updated (long ago) so this is no longer necessary.
We can return this token to valid namespace.

This has been in my tree for many months and has survived buddybuilds
and a variety of FastLED versions.