Merge my audio layer for ESP-IDF5 (Arduino3) and ESP-IDF4 (Arduino2)
It's my goal to land this before we switch, but to have the code in the
tree. It needs to be exercided on more hardware than I have right now
(it's fine on the analog mic of Mesmerizer) but it deserves widespread
testing.  (Gemini help me wrangle this code to "prove" it's right, so
I'll let it bigger part rest of this.)